This is a solution to the Sorted Squares problem under Arrays101 on Leetcode.
The code takes a sorted array of integers, squares them, and then resorts the array.
For all numbers >=0, the order would be the same. However, this program accounts for negative number.
For example the square of -4 has a larger value than the square of positive 3, so it should be sorted as greater.